moved SecurityHeaders + adjusted import in test
